+++ scripts/addons/cam/ui.py
@@ -33,9 +33,13 @@ from bpy.types import (Panel, Menu, Operator, PropertyGroup, )
 from cam import gcodeimportparser, simple
 from cam.simple import *
 
-from cam.ui_panels.info_panel import CAM_INFO_Panel
 from cam.ui_panels.buttons_panel import CAMButtonsPanel
-
+from cam.ui_panels.info import CAM_INFO_Panel
+from cam.ui_panels.operations import CAM_OPERATIONS_Panel
+from cam.ui_panels.cutter import CAM_CUTTER_Panel
+from cam.ui_panels.machine import CAM_MACHINE_Panel
+from cam.ui_panels.material import CAM_MATERIAL_Panel
+from cam.ui_panels.chains import CAM_UL_operations, CAM_UL_chains, CAM_CHAINS_Panel
 
 # Displays percentage of the cutter which is engaged with the material
 # Displays a warning for engagements greater than 50%
@@ -54,189 +58,6 @@ def EngagementDisplay(operat, layout):
         layout.label(text="Cutter Engagement: " + str(round(100 * ao.dist_between_paths / ao.cutter_diameter, 1)) + "%")
 
 
-class CAM_CUTTER_Panel(CAMButtonsPanel, bpy.types.Panel):
-    """CAM cutter panel"""
-    bl_label = " "
-    bl_idname = "WORLD_PT_CAM_CUTTER"
-
-    COMPAT_ENGINES = {'BLENDERCAM_RENDER'}
-
-    def draw_header(self, context):
-        self.layout.menu("CAM_CUTTER_MT_presets", text="CAM Cutter")
-
-    def draw(self, context):
-        layout = self.layout
-        d = bpy.context.scene
-        if len(d.cam_operations) > 0:
-            ao = d.cam_operations[d.cam_active_operation]
-
-            if ao:
-                # cutter preset
-                row = layout.row(align=True)
-                row.menu("CAM_CUTTER_MT_presets", text=bpy.types.CAM_CUTTER_MT_presets.bl_label)
-                row.operator("render.cam_preset_cutter_add", text="", icon='ADD')
-                row.operator("render.cam_preset_cutter_add", text="", icon='REMOVE').remove_active = True
-                layout.prop(ao, 'cutter_id')
-                layout.prop(ao, 'cutter_type')
-                if ao.cutter_type == 'VCARVE':
-                    layout.prop(ao, 'cutter_tip_angle')
-                if ao.cutter_type == 'BALLCONE':
-                    layout.prop(ao, 'ball_radius')
-                    EngagementDisplay(ao, layout)
-                    layout.prop(ao, 'cutter_tip_angle')
-                    layout.label(text='Cutter diameter = shank diameter')
-                if ao.cutter_type == 'CYLCONE':
-                    layout.prop(ao, 'cylcone_diameter')
-                    EngagementDisplay(ao, layout)
-                    layout.prop(ao, 'cutter_tip_angle')
-                    layout.label(text='Cutter diameter = shank diameter')
-                if ao.cutter_type == 'BULLNOSE':
-                    layout.prop(ao, 'bull_corner_radius')
-                    EngagementDisplay(ao, layout)
-                    layout.label(text='Cutter diameter = shank diameter')
-
-                if ao.cutter_type == 'LASER':
-                    layout.prop(ao, 'Laser_on')
-                    layout.prop(ao, 'Laser_off')
-                    layout.prop(ao, 'Laser_cmd')
-                    layout.prop(ao, 'Laser_delay')
-
-                if ao.cutter_type == 'PLASMA':
-                    layout.prop(ao, 'Plasma_on')
-                    layout.prop(ao, 'Plasma_off')
-                    layout.prop(ao, 'Plasma_delay')
-                    layout.prop(ao, 'Plasma_dwell')
-                    layout.prop(ao, 'lead_in')
-                    layout.prop(ao, 'lead_out')
-
-                if ao.cutter_type == 'CUSTOM':
-                    if ao.use_exact:
-                        layout.label(text='Warning - only convex shapes are supported. ', icon='COLOR_RED')
-                        layout.label(text='If your custom cutter is concave,')
-                        layout.label(text='switch exact mode off.')
-
-                    layout.prop_search(ao, "cutter_object_name", bpy.data, "objects")
-
-                layout.prop(ao, 'cutter_diameter')
-                if ao.strategy == "POCKET" or ao.strategy == "PARALLEL" or ao.strategy == "CROSS" \
-                        or ao.strategy == "WATERLINE":
-                    EngagementDisplay(ao, layout)
-                if ao.cutter_type != "LASER":
-                    layout.prop(ao, 'cutter_flutes')
-                layout.prop(ao, 'cutter_description')
-
-
-class CAM_MACHINE_Panel(CAMButtonsPanel, bpy.types.Panel):
-    """CAM machine panel"""
-    bl_label = " "
-    bl_idname = "WORLD_PT_CAM_MACHINE"
-
-    COMPAT_ENGINES = {'BLENDERCAM_RENDER'}
-
-    def draw_header(self, context):
-        self.layout.menu("CAM_MACHINE_MT_presets", text="CAM Machine")
-
-    def draw(self, context):
-        layout = self.layout
-        s = bpy.context.scene
-        us = s.unit_settings
-
-        ao = s.cam_machine
-
-        if ao:
-            use_experimental = bpy.context.preferences.addons['cam'].preferences.experimental
-
-            # machine preset
-            row = layout.row(align=True)
-            row.menu("CAM_MACHINE_MT_presets", text=bpy.types.CAM_MACHINE_MT_presets.bl_label)
-            row.operator("render.cam_preset_machine_add", text="", icon='ADD')
-            row.operator("render.cam_preset_machine_add", text="", icon='REMOVE').remove_active = True
-            layout.prop(ao, 'post_processor')
-            layout.prop(ao, 'eval_splitting')
-            if ao.eval_splitting:
-                layout.prop(ao, 'split_limit')
-
-            layout.prop(us, 'system')
-
-            layout.prop(ao, 'use_position_definitions')
-            if ao.use_position_definitions:
-                layout.prop(ao, 'starting_position')
-                layout.prop(ao, 'mtc_position')
-                layout.prop(ao, 'ending_position')
-            layout.prop(ao, 'working_area')
-            layout.prop(ao, 'feedrate_min')
-            layout.prop(ao, 'feedrate_max')
-            layout.prop(ao, 'feedrate_default')
-            # TODO: spindle default and feedrate default should become part of the cutter definition...
-            layout.prop(ao, 'spindle_min')
-            layout.prop(ao, 'spindle_max')
-            layout.prop(ao, 'spindle_start_time')
-            layout.prop(ao, 'spindle_default')
-            layout.prop(ao, 'output_tool_definitions')
-            layout.prop(ao, 'output_tool_change')
-            if ao.output_tool_change:
-                layout.prop(ao, 'output_g43_on_tool_change')
-
-            if use_experimental:
-                layout.prop(ao, 'axis4')
-                layout.prop(ao, 'axis5')
-                layout.prop(ao, 'collet_size')
-
-                layout.prop(ao, 'output_block_numbers')
-                if ao.output_block_numbers:
-                    layout.prop(ao, 'start_block_number')
-                    layout.prop(ao, 'block_number_increment')
-            layout.prop(ao, 'hourly_rate')
-
-
-class CAM_MATERIAL_Panel(CAMButtonsPanel, bpy.types.Panel):
-    """CAM material panel"""
-    bl_label = "CAM Material size and position"
-    bl_idname = "WORLD_PT_CAM_MATERIAL"
-
-    COMPAT_ENGINES = {'BLENDERCAM_RENDER'}
-
-    def draw(self, context):
-        layout = self.layout
-        scene = bpy.context.scene
-
-        if len(scene.cam_operations) == 0:
-            layout.label(text='Add operation first')
-        if len(scene.cam_operations) > 0:
-            ao = scene.cam_operations[scene.cam_active_operation]
-            if ao:
-                layout.template_running_jobs()
-                if ao.geometry_source in ['OBJECT', 'COLLECTION']:
-                    layout.prop(ao, 'material_from_model')
-
-                    if ao.material_from_model:
-                        layout.prop(ao, 'material_radius_around_model')
-                    else:
-                        layout.prop(ao, 'material_origin')
-                        layout.prop(ao, 'material_size')
-
-                    layout.prop(ao, 'material_center_x')
-                    layout.prop(ao, 'material_center_y')
-                    layout.prop(ao, 'material_Z')
-                    layout.operator("object.cam_position", text="Position object")
-                else:
-                    layout.label(text='Estimated from image')
-
-
-class CAM_UL_operations(UIList):
-    def draw_item(self, context, layout, data, item, icon, active_data, active_propname, index):
-        operation = item
-        if self.layout_type in {'DEFAULT', 'COMPACT'}:
-
-            layout.label(text=item.name, translate=False, icon_value=icon)
-            icon = 'LOCKED' if operation.computing else 'UNLOCKED'
-            if operation.computing:
-                layout.label(text=operation.outtext)  # "computing" )
-        elif self.layout_type in {'GRID'}:
-            layout.alignment = 'CENTER'
-            layout.label(text="", icon_value=icon)
-
-
 class CAM_UL_orientations(UIList):
     def draw_item(self, context, layout, data, item, icon, active_data, active_propname, index):
         if self.layout_type in {'DEFAULT', 'COMPACT'}:
@@ -245,156 +66,6 @@ class CAM_UL_orientations(UIList):
         elif self.layout_type in {'GRID'}:
             layout.alignment = 'CENTER'
             layout.label(text="", icon_value=icon)
-
-
-class CAM_UL_chains(UIList):
-    def draw_item(self, context, layout, data, item, icon, active_data, active_propname, index):
-        chain = item
-        if self.layout_type in {'DEFAULT', 'COMPACT'}:
-
-            layout.label(text=item.name, translate=False, icon_value=icon)
-            icon = 'LOCKED' if chain.computing else 'UNLOCKED'
-            if chain.computing:
-                layout.label(text="computing")
-        elif self.layout_type in {'GRID'}:
-            layout.alignment = 'CENTER'
-            layout.label(text="", icon_value=icon)
-
-
-class CAM_CHAINS_Panel(CAMButtonsPanel, bpy.types.Panel):
-    """CAM chains panel"""
-    bl_label = "CAM chains"
-    bl_idname = "WORLD_PT_CAM_CHAINS"
-
-    COMPAT_ENGINES = {'BLENDERCAM_RENDER'}
-
-    def draw(self, context):
-        layout = self.layout
-
-        row = layout.row()
-        scene = bpy.context.scene
-
-        row.template_list("CAM_UL_chains", '', scene, "cam_chains", scene, 'cam_active_chain')
-        col = row.column(align=True)
-        col.operator("scene.cam_chain_add", icon='ADD', text="")
-        col.operator("scene.cam_chain_remove", icon='REMOVE', text="")
-
-        if len(scene.cam_chains) > 0:
-            chain = scene.cam_chains[scene.cam_active_chain]
-            row = layout.row(align=True)
-
-            if chain:
-                row.template_list("CAM_UL_operations", '', chain, "operations", chain, 'active_operation')
-                col = row.column(align=True)
-                col.operator("scene.cam_chain_operation_add", icon='ADD', text="")
-                col.operator("scene.cam_chain_operation_remove", icon='REMOVE', text="")
-                if len(chain.operations) > 0:
-                    col.operator("scene.cam_chain_operation_up", icon='TRIA_UP', text="")
-                    col.operator("scene.cam_chain_operation_down", icon='TRIA_DOWN', text="")
-
-                if not chain.computing:
-                    if chain.valid:
-                        pass
-                        layout.operator("object.calculate_cam_paths_chain", text="Calculate chain paths & Export Gcode")
-                        layout.operator("object.cam_export_paths_chain", text="Export chain gcode")
-                        layout.operator("object.cam_simulate_chain", text="Simulate this chain")
-                    else:
-                        layout.label(text="chain invalid, can't compute")
-                else:
-                    layout.label(text='chain is currently computing')
-
-                layout.prop(chain, 'name')
-                layout.prop(chain, 'filename')
-
-
-class CAM_OPERATIONS_Panel(CAMButtonsPanel, bpy.types.Panel):
-    """CAM operations panel"""
-    bl_label = "CAM operations"
-    bl_idname = "WORLD_PT_CAM_OPERATIONS"
-
-    COMPAT_ENGINES = {'BLENDERCAM_RENDER'}
-
-    def draw(self, context):
-        layout = self.layout
-
-        row = layout.row()
-        scene = bpy.context.scene
-        row.template_list("CAM_UL_operations", '', scene, "cam_operations", scene, 'cam_active_operation')
-        col = row.column(align=True)
-        col.operator("scene.cam_operation_add", icon='ADD', text="")
-        col.operator("scene.cam_operation_copy", icon='COPYDOWN', text="")
-        col.operator("scene.cam_operation_remove", icon='REMOVE', text="")
-        col.separator()
-        col.operator("scene.cam_operation_move", icon='TRIA_UP', text="").direction = 'UP'
-        col.operator("scene.cam_operation_move", icon='TRIA_DOWN', text="").direction = 'DOWN'
-
-        if len(scene.cam_operations) > 0:
-            use_experimental = bpy.context.preferences.addons['cam'].preferences.experimental
-            ao = scene.cam_operations[scene.cam_active_operation]
-
-            row = layout.row(align=True)
-            row.menu("CAM_OPERATION_MT_presets", text=bpy.types.CAM_OPERATION_MT_presets.bl_label)
-            row.operator("render.cam_preset_operation_add", text="", icon='ADD')
-            row.operator("render.cam_preset_operation_add", text="", icon='REMOVE').remove_active = True
-
-            if ao:
-                if not ao.computing:
-                    if ao.valid:
-                        layout.operator("object.calculate_cam_path", text="Calculate path & export Gcode")
-                        if ao.name is not None:
-                            name = "cam_path_{}".format(ao.name)
-                            if scene.objects.get(name) is not None:
-                                layout.operator("object.cam_export", text="Export Gcode ")
-                        layout.operator("object.cam_simulate", text="Simulate this operation")
-
-
-                    else:
-                        layout.label(text="operation invalid, can't compute")
-                else:
-                    row = layout.row(align=True)
-                    row.label(text='computing')
-                    row.operator('object.kill_calculate_cam_paths_background', text="", icon='CANCEL')
-
-                sub = layout.column()
-                sub.active = not ao.computing
-
-                sub.prop(ao, 'name')
-                sub.prop(ao, 'filename')
-
-                layout.prop(ao, 'geometry_source')
-                if not ao.strategy == 'CURVE':
-                    if ao.geometry_source == 'OBJECT':
-                        layout.prop_search(ao, "object_name", bpy.data, "objects")
-                        if ao.enable_A:
-                            layout.prop(ao, 'rotation_A')
-                        if ao.enable_B:
-                            layout.prop(ao, 'rotation_B')
-
-                    elif ao.geometry_source == 'COLLECTION':
-                        layout.prop_search(ao, "collection_name", bpy.data, "collections")
-                    else:
-                        layout.prop_search(ao, "source_image_name", bpy.data, "images")
-                else:
-                    if ao.geometry_source == 'OBJECT':
-                        layout.prop_search(ao, "object_name", bpy.data, "objects")
-                    elif ao.geometry_source == 'COLLECTION':
-                        layout.prop_search(ao, "collection_name", bpy.data, "collections")
-
-                if ao.strategy in ['CARVE', 'PROJECTED_CURVE']:
-                    layout.prop_search(ao, "curve_object", bpy.data, "objects")
-                    if ao.strategy == 'PROJECTED_CURVE':
-                        layout.prop_search(ao, "curve_object1", bpy.data, "objects")
-
-                layout.prop(ao, 'remove_redundant_points')
-                if ao.remove_redundant_points:
-                    layout.label(text='Revise your Code before running!')
-                    layout.label(text='Quality will suffer if tolerance')
-                    layout.label(text='is high')
-                    layout.prop(ao, 'simplify_tol')
-                if ao.geometry_source in ['OBJECT', 'COLLECTION']:
-                    layout.prop(ao, 'use_modifiers')
-                layout.prop(ao, 'hide_all_others')
-                layout.prop(ao, 'parent_path_to_object')
